refactor countries hub
