+++ app/assets/scripts/views/countries-hub/index.js
@@ -1,0 +1,3 @@
+import CountriesHub from './countries-hub';
+
+export default CountriesHub;
